Fix incorrect value in Mekanism tools config
The default value of `refined_obsidianHoeDamage` in Mekanism Tools' `tools-materials-startup.toml` config file is incompatible with the value of `refined_obsidianAttackDamage` in such a way that the automatic fixer which runs when invalid config values are detected tries to run repeatedly to no effect while filling up the log files.

I've changed the `refined_obsidianHoeDamage` value to `-5.0`, which is the lowest supported value given the value of `refined_obsidianAttackDamage`. This should fix the repeating error.