Replicator II, sounds for some EI machines.
Replicator II is an electric singleblock, which allows it to be overclocked, it also doesn't copy NBT, which makes us able to be more lenient on its blacklist. Recipe for it is TBD.

Added sounds to 5 EI machines, Bending Machine, Alloy Smelter, Canning Machine, Composter and Brewery. Would've added to all, but due to the way the sound addon works, I wasn't able to, will look into it later.