Add Particle Accelerator multiblock - W.I.P
Introduces a particle accelerator multiblock, which handles Fusion Reactor recipes, but is capable of doing them in batches of up-to 128, with a 50% discount on energy cost. Commented out for now.

No recipe or texture for the multiblock as of yet. Structure of it is still to be finalized.

Extends custom MI machine definitions to include a 'mod' property, allowing machines to be associated with mods other than 'modern_industrialization'.

Updates asset generation, language entry creation, and loot table generation to utilize the 'mod' property when present.