Balance Just Dire Things fluid fuel values
This adjusts the fuel values, in EU/mB, of the three fluid fuels from Just Dire Things to fully take into account the fuel value of the JDT coal corresponding to each fluid tier. This makes it practical to produce JDT fluid fuels even after the reduction in JDT coal enrichment yields in the Ooze Lab from 1:1 to 9:4.